fixed auth interceptor

+++ code/jvm/src/main/kotlin/daw/isel/pt/gomoku/controllers/pipeline/AuthInterceptor.kt
@@ -17,7 +17,7 @@ class AuthInterceptor(private val tokenValidator: RequestTokenValidator): Handle
             throw UnauthorizedException("Unauthorized Access")
         }
         logger.info("Handling Auth interceptor")
-        val userId = request.requestURI.split("/")[2].toInt()
+        val userId = request.requestURI.split("/")[2]
         val authenticatedUser = tokenValidator.processAuthorizationHeaderValue(authorizationHeader, userId)
         return if(authenticatedUser == null) throw UnauthorizedException("Unauthorized Access")
         else true
